list -> longtensor in eval

+++ kbc/models.py
@@ -62,9 +62,9 @@ class KBCModel(nn.Module, ABC):
                                 int(x - c_begin) for x in filter_out
                                 if c_begin <= x < c_begin + chunk_size
                             ]
-                            scores[i, filter_in_chunk] = -1e6
+                            scores[i, torch.LongTensor(filter_in_chunk)] = -1e6
                         else:
-                            scores[i, filter_out] = -1e6
+                            scores[i, torch.LongTensor(filter_out)] = -1e6
                     ranks[b_begin:b_begin + batch_size] += torch.sum(
                         (scores >= targets).float(), dim=1
                     ).cpu()
